Whitted Shader with Reflections and dynamic selection of Integrators

+++ src/Core/TriMeshImp.cpp
@@ -20,7 +20,7 @@ TriMeshImp::~TriMeshImp()
 int TriMeshImp::PushVertex(const Vector3& location)
 {
 	_vertexLocations.push_back(location);
-	return _vertexLocations.size() -1;
+	return (int)(_vertexLocations.size() -1);
 }
 
 int TriMeshImp::PushTriangle(int vertex1,int vertex2,int vertex3,Material material)
@@ -29,18 +29,18 @@ int TriMeshImp::PushTriangle(int vertex1,int vertex2,int vertex3,Material materi
 	for(auto it = _materials.begin(); it != _materials.end(); ++it)
 		if(*it == material)
 		{
-			material_index = it - _materials.begin();
+			material_index = (int)(it - _materials.begin());
 			break;
 		}
 	if(material_index == -1)
 	{
-		material_index = _materials.size();
+		material_index = (int)(_materials.size());
 		_materials.push_back(material);
 	}
 
 	_triangleVertices.push_back(Vector3i(vertex1,vertex2,vertex3));
 	_triangleMaterials.push_back(material_index);
-	return _triangleVertices.size() - 1;
+	return (int)(_triangleVertices.size() - 1);
 }
 
 RObjectType ITriMesh::ObjectType = ObjectType::TriMesh;
